net: socket: Implement exponentially growing polling delay in socket send
In both zsock_sendto_ctx and zsock_sendmsg_ctx, when the send cannot
complete due to either a failure for buffer allocation or full socket
send buffer, blocking is implemented as a polling loop.

This polling loop used a fixed 100 ms delay. This can be conservative
in most situations resulting in slow throughput. Lowering this delay
can cause a very high system load, which does not scale.

Using exponentially increasing delays, is a happy medium where in
the positive case can use a quick response and grows quickly to prevent
loading the system too much when the polling fails for longer time.

Signed-off-by: Sjors Hettinga <s.a.hettinga@gmail.com>