net: tcp: Basic receive window update functionality
When the host sends data till the Zephyr receive window becomes full,
the host stops sending until it hits a timer to send a zero window
probe message.

An basic implementation of the receive window update. When there is a
significant amount of room available in the receive buffer, actively
notify the host. This will prevent the link from stalling for a long
time.

TODO:
- Make it work using the SO_RCVBUF socket option
- Avoid sending out small receive window advertizements
    (Silly window update)

Signed-off-by: Sjors Hettinga <s.a.hettinga@gmail.com>